requirements.txt
numpy==1.22.1
opencv-python==4.5.5.62
progressbar2==4.0.0
python-utils==3.1.0
